refactor(package): change packages
- 즐겨찾기 도메인 설계 변경에 따른 패키지 분리
