+++ src/main/java/nextstep/subway/favorite/ui/dto/StationInFavoriteResponse.java
@@ -1,4 +1,4 @@
-package nextstep.subway.member.dto;
+package nextstep.subway.favorite.ui.dto;
 
 import java.time.LocalDateTime;
 import java.util.Objects;
